requirements.txt
appdirs==1.4.4
asgiref==3.4.1
audioread==2.1.9
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
decorator==5.0.9
Django==3.2.6
django-cors-headers==3.7.0
idna==3.2
imbalanced-learn==0.8.0
imblearn==0.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
librosa==0.8.1
lightgbm==3.2.1
llvmlite==0.37.0
MarkupSafe==2.0.1
numba==0.54.0
numpy==1.20.3
packaging==21.0
pandas==1.3.2
pooch==1.5.1
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
resampy==0.2.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
SoundFile==0.10.3.post1
sqlparse==0.4.1
threadpoolctl==2.2.0
urllib3==1.26.6
Werkzeug==2.0.1
tensorflow-gpu==2.9.1
opencv==3.4.1
matplotlib
